Moved package + added build bazel
