+++ cmd/agent/subcommands/run/command_observer.go
@@ -15,9 +15,9 @@ package run
 import (
 	"go.uber.org/fx"
 
+	hfrunnerfx "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/fx"
 	logssourcefx "github.com/DataDog/datadog-agent/comp/anomalydetection/logssource/fx"
 	observerfx "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/fx"
-	hfrunnerfx "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/hfrunner/fx"
 	recorderfx "github.com/DataDog/datadog-agent/comp/anomalydetection/recorder/fx-noop"
 	reporterfx "github.com/DataDog/datadog-agent/comp/anomalydetection/reporter/fx"
 )
