+++ comp/anomalydetection/hfrunner/def/BUILD.bazel
@@ -3,10 +3,10 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "def",
     srcs = ["component.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/hfrunner/def",
+    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/def",
     visibility = ["//visibility:public"],
     deps = [
-        "//comp/anomalydetection/observer/def",
         "//pkg/metrics",
+        "@com_github_datadog_datadog_agent_comp_anomalydetection_observer_def//:def",
     ],
 )
