+++ comp/anomalydetection/hfrunner/def/component.go
