+++ comp/anomalydetection/hfrunner/impl/BUILD.bazel
@@ -1,0 +1,41 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "impl",
+    srcs = [
+        "hfrunner.go",
+        "runner.go",
+        "sender.go",
+    ],
+    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/impl",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//comp/anomalydetection/hfrunner/def",
+        "//comp/core/config",
+        "//comp/core/tagger/def",
+        "//comp/core/workloadfilter/def",
+        "//comp/core/workloadmeta/def",
+        "//comp/def",
+        "//pkg/metrics",
+        "//pkg/metrics/event",
+        "//pkg/metrics/servicecheck",
+        "//pkg/serializer/types",
+        "//pkg/util/log",
+        "//pkg/util/option",
+        "@//comp/core/autodiscovery/integration",
+        "@//pkg/aggregator/sender",
+        "@//pkg/collector/check",
+        "@//pkg/collector/check/id",
+        "@//pkg/collector/check/stats",
+        "@//pkg/collector/corechecks/containers/generic",
+        "@//pkg/collector/corechecks/net/network",
+        "@//pkg/collector/corechecks/system/cpu/cpu",
+        "@//pkg/collector/corechecks/system/cpu/load",
+        "@//pkg/collector/corechecks/system/disk/disk",
+        "@//pkg/collector/corechecks/system/disk/io",
+        "@//pkg/collector/corechecks/system/filehandles",
+        "@//pkg/collector/corechecks/system/memory",
+        "@//pkg/collector/corechecks/system/uptime",
+        "@com_github_datadog_datadog_agent_comp_anomalydetection_observer_def//:def",
+    ],
+)
